app/requirements.txt
fastapi
uvicorn[standard]
pillow
